+++ orchestrate/orchestrate/api/v1/api.go
@@ -20,16 +20,11 @@ func BindAPI(router *gin.Engine, authMiddleware *middleware.GinJWTMiddleware) er
 	admin := v1.Group("/admin")
 	admin.Use(authMiddleware.MiddlewareFunc(middleware.AdminAuthorizor))
 	{
-		admin.GET("/jobs", handler.ListJobsHandler)
-		admin.POST("/job", handler.AddJobHandler)
-		admin.DELETE("/job/:job_id", handler.DeleteJobHandler)
-	}
+		admin.GET("/alerts", handler.ListAlertsHandler)
+		admin.POST("/alert", handler.AddAlertHandler)
+		admin.DELETE("/alert/:alert_id", handler.DeleteAlertHandler)
 
-	orchestration := v1.Group("/")
-	// This means that authentication is optional
-	orchestration.Use(authMiddleware.MiddlewareFunc(middleware.NullAuthorizor))
-	{
-		orchestration.POST("/experiment", handler.AddExperimentHandler)
+		admin.POST("/experiment", handler.AddExperimentHandler)
 	}
 
 	rendezvous := v1.Group("/")
@@ -45,8 +40,15 @@ func BindAPI(router *gin.Engine, authMiddleware *middleware.GinJWTMiddleware) er
 	device := v1.Group("/")
 	device.Use(authMiddleware.MiddlewareFunc(middleware.DeviceAuthorizor))
 	{
-		device.GET("/tasks", handler.ListTasksHandler)
+		/*
+			experiments.GET("/experiments", handler.ListExperimentsHandler)
+			experiments.GET("/experiment/:exp_id", handler.GetExperimentHandler)
+			experiments.GET("/experiment/:exp_id/accept", handler.AcceptExperimentHandler)
+			experiments.GET("/experiment/:exp_id/reject", handler.RejectExperimentHandler)
+			experiments.GET("/experiment/:exp_id/done", handler.DoneExperimentHandler)
+		*/
 
+		device.GET("/tasks", handler.ListTasksHandler)
 		device.GET("/task/:task_id", handler.GetTaskHandler)
 		device.POST("/task/:task_id/accept", handler.AcceptTaskHandler)
 		device.POST("/task/:task_id/reject", handler.RejectTaskHandler)
