+++ orchestrate/orchestrate/sched/utils_test.go
@@ -17,10 +17,10 @@ func TestParseSchedule(t *testing.T) {
 		t.Errorf("expected Repeat to be 42 (got: %d)", s.Repeat)
 	}
 	if s.Duration.Weeks != 1.3 {
-		t.Error("expected 1.3 weeks duration (got: %f)", s.Duration.Weeks)
+		t.Errorf("expected 1.3 weeks duration (got: %f)", s.Duration.Weeks)
 	}
 	if s.Duration.Minutes != 2.0 {
-		t.Error("expected 2.0 minutes duration (got: %f)", s.Duration.Minutes)
+		t.Errorf("expected 2.0 minutes duration (got: %f)", s.Duration.Minutes)
 	}
 
 	s, err = ParseSchedule("R/2018-12-16T16:20:30Z/PT2M")
